tests/wpt/web-platform-tests/dom/nodes/Element-hasAttribute.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Element.prototype.hasAttribute</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-element-hasattribute">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<span data-e2="2" data-F2="3" id="t"></span>

<script>
"use strict";

test(() => {

  const el = document.createElement("p");
  el.setAttributeNS("foo", "x", "first");

  assert_true(el.hasAttribute("x"));

}, "hasAttribute should check for attribute presence, irrespective of namespace");

test(() => {

  const el = document.getElementById("t");

  assert_true(el.hasAttribute("data-e2"));
  assert_true(el.hasAttribute("data-E2"));
  assert_true(el.hasAttribute("data-f2"));
  assert_true(el.hasAttribute("data-F2"));

}, "hasAttribute should work with all attribute casings");
</script>
